rename annotation plugin
