+++ emgapianns/apps.py
@@ -14,4 +14,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-default_app_config = 'emgapimetadata.apps.EmgapimetadataConfig'
+from django.apps import AppConfig
+
+
+class EmgapiannsConfig(AppConfig):
+    name = 'emgapianns'
